requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
shap==0.44.0
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.31.0
plotly==5.18.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.0
requests==2.31.0
joblib==1.3.2
imbalanced-learn==0.11.0
great-expectations==0.18.8
dbt-postgres==1.7.4
pytest==7.4.4
httpx==0.26.0
